Update README.md
Basically we will gives the instructions to the user based on their commands . user has to enter to start,stop and quit,help and if something entered other than these we display the user that i don't understand. when help command enter by user we have to give instructions that what start and stop ,quit commands meant to be.